pulled the proxy server impl into the server abstraction
